Move utils into dedicated folders
