+++ utils/ajax.js
@@ -1,0 +1,12 @@
+export function getScript(src) {
+  const p = new Promise((resolve, reject) => {
+    const script = document.createElement("script");
+    document.body.appendChild(script);
+    script.onload = resolve;
+    script.onerror = reject;
+    script.async = true;
+    script.src = src;
+  });
+
+  return p;
+}
